Decimal: change default operation modes for SetFloat and SetFloat64.
This will be the last change to this: the conversion in Float and
SetFloat64 is performed using the receiver's precision and rounding
mode.

If z's precision is larger than the precision of the Float to be
converted, this may lead to inconsistencies between the outputs of
x.Text and z.Text. A good example of this is NewFloat(1.1): this
is not an exact number in base 2. If converting to Decimal with a
precision higher than 16, the result will be around 1.1000000000000001
which is not desirable in most cases.

There are however cases where one might want to use Float for fast
base-2 computations, then convert to Decimal with high precision.
Also considering that all other functions operate using the receiver's
precision and rounding mode, SetFloat and SetFloat64 have been changed
to use this convention as well.

The SetFloat doc has been updated to include an example showing how to
preserve z.Text == x.Text.