dec/arith: boost shl10VU/shr10VU performance
These functions now use the "constant division by multiplication"
algorithm. This is made possible by the small range of possible values
for the shift count (18 at most) and a precoomputed table of magic
numbers (uses the same algorithm as the Go compiler). The amd64
assembler version of shl10VU now has a slight performance gain over the
go version (about 5% less time per op).

The performance boost is significative:

    name     old time/op  new time/op  delta
    Shl10VU  9.06µs ± 0%  3.32µs ± 2%  -63.36%  (p=0.008 n=5+5)

This improves (*Decimal).Add/Sub performance as well (when the mantissae
of the operands need to be aligned) and Decimal -> Int(64) conversions.