math/exp: tweak exp core to handle both e^x and e^x-1
Moved back most of the exponential algorithm to expT and have it handle
both e^x and e^x-1. This was to prevent an unnecessary triple precision
for Exp when |x| close to 0 and code duplicatioon just to handle this.
Exp and Expm1 do the special cases and bounds checking, then hand over
the rest of the computation to expT.

Now the internal working precision for Exp is z.Prec()*2 in the worst
case.